LP_SOLVER/requirements.txt
flask
gunicorn
google-generativeai
numpy
scipy